src/image_processor.py
"""
画像処理・OCRモジュール
"""

import os
import base64
from pathlib import Path
from typing import Dict, Any, Optional
import logging
from PIL import Image
import io

# Google Cloud Vision API
try:
    from google.cloud import vision
    from google.cloud.vision_v1 import types
except ImportError:
    vision = None
    logging.warning("Google Cloud Vision APIが利用できません")

class ImageProcessor:
    """画像処理・OCRクラス"""
    
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.logger = logging.getLogger(__name__)
        
        # Google Cloud Vision APIの初期化
        if vision:
            try:
                # 認証情報が設定されているかチェック
                import os
                if os.getenv("GOOGLE_APPLICATION_CREDENTIALS") or os.getenv("GOOGLE_CLOUD_PROJECT_ID"):
                    self.vision_client = vision.ImageAnnotatorClient()
                    self.logger.info("Google Cloud Vision APIを初期化しました")
                else:
                    self.logger.warning("Google Cloud認証情報が設定されていないため、ダミーモードで動作します")
                    self.vision_client = None
            except Exception as e:
                self.logger.error(f"Google Cloud Vision APIの初期化に失敗: {str(e)}")
                self.vision_client = None
        else:
            self.vision_client = None
    
    def load_image(self, receipt_id: str) -> Dict[str, Any]:
        """レシート画像を読み込み"""
        try:
            # 設定からファイルパターンを取得
            file_pattern = self.config["processing_flow"][0]["steps"][0]["target_file_pattern"]
            file_path = file_pattern.replace("{{receipt_id}}", receipt_id)
            
            image_path = Path(file_path)
            
            if not image_path.exists():
                # ダミー画像データを返す（テスト用）
                self.logger.warning(f"画像ファイルが見つかりません: {image_path}")
                return {
                    "image_loaded": True,
                    "is_dummy": True,
                    "message": "ダミー画像データを使用しています"
                }
            
            # 画像を読み込み
            with open(image_path, 'rb') as f:
                image_data = f.read()
            
            self.logger.info(f"画像を読み込みました: {image_path}")
            
            return {
                "image_loaded": True,
                "image_path": str(image_path),
                "image_size": len(image_data),
                "is_dummy": False
            }
            
        except Exception as e:
            self.logger.error(f"画像の読み込みに失敗しました: {str(e)}")
            raise
    
    def extract_text(self, receipt_id: str) -> Dict[str, Any]:
        """画像からテキストを抽出（OCR）"""
        try:
            if not self.vision_client:
                # ダミーテキストを返す（テスト用）
                self.logger.warning("Google Cloud Vision APIが利用できないため、ダミーテキストを使用します")
                return self._get_dummy_text()
            
            # 画像を読み込み
            image_data = self.load_image(receipt_id)
            
            if image_data.get("is_dummy", False):
                return self._get_dummy_text()
            
            # Google Cloud Vision APIでテキスト抽出
            image_path = image_data["image_path"]
            
            with open(image_path, 'rb') as f:
                content = f.read()
            
            image = types.Image(content=content)
            
            # テキスト検出
            response = self.vision_client.text_detection(image=image)
            texts = response.text_annotations
            
            if texts:
                extracted_text = texts[0].description
                self.logger.info(f"テキスト抽出成功: {len(extracted_text)}文字")
                
                return {
                    "extracted_text": extracted_text,
                    "text_length": len(extracted_text),
                    "confidence": "high",
                    "is_dummy": False
                }
            else:
                self.logger.warning("テキストが検出されませんでした")
                return self._get_dummy_text()
                
        except Exception as e:
            self.logger.error(f"テキスト抽出に失敗しました: {str(e)}")
            # エラー時はダミーテキストを返す
            return self._get_dummy_text()
    
    def _get_dummy_text(self) -> Dict[str, Any]:
        """ダミーテキストを返す（テスト用）"""
        # 実際のRED ELEPHANTレシートの情報を使用
        red_elephant_text = """RED ELEPHANT
123 Jalan Bukit Bintang
Kuala Lumpur, Malaysia

Kao Mun Gai Chicken White Rice
RM 15.90

Tax: RM 0.00
Total: RM 15.90

Thank you for your visit!
Date: 2024-01-15
Time: 14:30"""
        
        return {
            "extracted_text": red_elephant_text,
            "text_length": len(red_elephant_text),
            "confidence": "dummy",
            "is_dummy": True,
            "message": "RED ELEPHANTレシートのダミーテキストを使用しています"
        }
